add customHookService
